Tests/LibWeb/Text/expected/WebAnimations/misc/animating-all.txt

opacity value: 0.5
